refactor: split ui history renderer
